requirements.txt
# ── Core ──────────────────────────────────────
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.20
httpx==0.28.1

# ── LLM APIs ──────────────────────────────────
openai==1.59.4
google-genai==1.5.0 

# ── Embeddings & Reranking ─────────────────────
sentence-transformers==3.3.1
torch

# ── Vector DB ─────────────────────────────────
chromadb==0.6.3

# ── BM25 ──────────────────────────────────────
rank-bm25==0.2.2

# ── PDF ───────────────────────────────────────
pypdf==5.1.0

# ── Memory ────────────────────────────────────
redis[asyncio]==5.2.1

# ── Database / SQL ────────────────────────────
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
asyncpg==0.30.0

# ── Validation & Config ───────────────────────
pydantic==2.10.3
pydantic-settings==2.6.1

# ── Utilities ─────────────────────────────────
python-dotenv==1.0.1
tiktoken==0.8.0
numpy==2.1.3
nltk==3.9.1
